# ProductTraceability_2026/doc/openapi.yaml
openapi: 3.0.3
info:
  title: Agricultural Traceability System API
  version: 1.0.0
  description: API specification for the agricultural product traceability system.
servers:
  - url: '{baseUrl}'
    variables:
      baseUrl:
        default: http://localhost:8888
paths:
...
  /api/processing/create:
    post:
      summary: "[Full] Create processing record - resolves sourceBatchCode to batchId, auto-fills processingEnterpriseId, uploads hash to blockchain"
      tags: [Processing]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TraceProcessing'
      responses:
        '200':
          description: Processing record created, batch status updated to 5, dataHash + txHash populated
  /api/processing/list:
    get:
      summary: "[Full] List processing records with batch info enrichment"
      description: |
        Returns paginated processing records enriched with batch info (batchCode, productName, breed, unit, batchStatus).
        Supports enterpriseId and keyword filtering.
      tags: [Processing]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: pageNum, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 10 } }
        - { name: batchId, in: query, required: false, schema: { type: integer } }
        - { name: enterpriseId, in: query, required: false, schema: { type: integer } }
        - { name: keyword, in: query, required: false, schema: { type: string }, description: "Search by sourceBatchCode, processMethod, or operator" }
      responses:
        '200':
          description: OK

  # ==========================================
  # Storage
  # ==========================================
  /api/storage/list:
    get:
      summary: "[Full] Query storage records with pagination"
      tags: [Storage]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: pageNum, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 10 } }
        - { name: batchId, in: query, schema: { type: integer } }
        - { name: storageType, in: query, schema: { type: integer, description: "1-inbound, 2-outbound, 3-inventory" } }
        - { name: keyword, in: query, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer }
                  data:
                    type: object
                    properties:
                      records: { type: array, items: { $ref: '#/components/schemas/TraceStorage' } }
                      total: { type: integer }
  /api/storage/create:
    post:
      summary: "[Full] Create a storage record - auto-fills storageUnit from batch"
      tags: [Storage]
      security: [{ BearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TraceStorage' }
      responses:
        '200':
          description: OK
  /api/storage/delete/{id}:
    delete:
      summary: "[Full] Delete a storage record"
      tags: [Storage]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: OK

  # ==========================================
  # Transport
  # ==========================================
...
  /api/system/log/operation:
    get:
      summary: "✅ Operation log list"
      tags: [System-Log]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: current, in: query, schema: { type: integer, default: 1 } }
        - { name: size, in: query, schema: { type: integer, default: 10 } }
        - { name: username, in: query, schema: { type: string } }
        - { name: operation, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: integer } }
        - { name: startDate, in: query, schema: { type: string, format: date } }
        - { name: endDate, in: query, schema: { type: string, format: date } }
      responses:
        '200':
          description: Paginated operation log list
  /api/system/log/login:
    get:
      summary: "✅ Login log list"
      tags: [System-Log]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: current, in: query, schema: { type: integer, default: 1 } }
        - { name: size, in: query, schema: { type: integer, default: 10 } }
        - { name: username, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: integer } }
        - { name: startDate, in: query, schema: { type: string, format: date } }
        - { name: endDate, in: query, schema: { type: string, format: date } }
      responses:
        '200':
          description: Paginated login log list

  # ==========================================
  # System Config
  # ==========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TraceProcessing:
      type: object
      properties:
        id: { type: integer }
        batchId: { type: integer, description: "Auto-resolved from sourceBatchCode" }
        sourceBatchCode: { type: string }
        processingEnterpriseId: { type: integer, description: "Auto-mapped from frontend enterpriseId" }
        processingDate: { type: string, format: date }
        processMethod: { type: string, description: "e.g. zhengji/fenqie/qugu" }
        specs: { type: string, description: "Packaging spec e.g. 500g/bag" }
        operator: { type: string }
        inputQuantity: { type: number }
        outputQuantity: { type: number }
        images: { type: string, description: "JSON array of image URLs" }
        dataHash: { type: string, description: "SHA-256 hash for blockchain" }
        txHash: { type: string }
        blockNumber: { type: integer }
        chainTime: { type: string, format: date-time }
        remark: { type: string }
    TraceStorage:
      type: object
      properties:
        id: { type: integer }
        batchId: { type: integer }
        batchCode: { type: string, description: "Frontend sends batchCode, backend resolves to batchId" }
        storageEnterpriseId: { type: integer }
        storageType: { type: integer, description: "1-inbound, 2-outbound, 3-inventory check" }
        storageDate: { type: string, format: date }
        warehouseName: { type: string, description: "Frontend field: warehouseName (was warehouseNo)" }
        warehouseLocation: { type: string }
        storageQuantity: { type: number }
        storageUnit: { type: string }
        temperature: { type: number, description: "Celsius" }
        humidity: { type: number, description: "Percentage" }
        storageCondition: { type: string }
        operator: { type: string, description: "Frontend field: operator (was manager)" }
        images: { type: string, description: "JSON array of image URLs" }
        txHash: { type: string }
        blockNumber: { type: integer }
        chainTime: { type: string, format: date-time }
        dataHash: { type: string }
        remark: { type: string }
...